requirements.txt
Django==4.2.5
Pillow==10.0.1
django-extensions==3.2.3
django-debug-toolbar==4.2.0
djangorestframework==3.14.0
httpie==3.2.2